3/requirements.txt
coreapi==2.2.4
coreschema==0.0.4
Django==1.10
django-cors-headers==2.2.0
git+git://github.com/Austyns/django-cors-headers.git
django-cors-middleware==1.2.0
django-extra-fields==0.8
django-filter==1.0.4
djangorestframework==3.5.4
gunicorn==19.8.1
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
openapi-codec==1.3.1
Pillow==3.3.0
Pygments==2.1.3
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
whitenoise==3.3.1
django-rest-swagger==2.1.1